% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghqmmfunctions.R
\name{aghqgamm}
\alias{aghqgamm}
\title{Binary Generalized Additive Mixed Model via AGHQ with Exact Gradients}
\usage{
aghqgamm(formula, data, k = 5, control = aghqmm_control())
}
\arguments{
\item{data}{A \code{data.frame} containing the variables found in \code{formula}.}

\item{k}{Order of the adaptive quadrature. In 1 dimension, \code{k} is the number of
points; in d-dimensions there are \code{k^d} total points.}

\item{control}{Output of \code{aghqmm_control()}, see that function.}

\item{smoothformula}{An R formula of the form \code{y ~ s(x, bs = "bs", k = 20)} compatible with
\code{mgcv::gam}; see that function for details.}

\item{glmmformula}{An R mixed model formula pf the form \code{y ~ x*t + (t|id)},
compatible with \code{lme4::glmer}; see that function for details.}
}
\value{
A list containing elements: TODO
}
\description{
Fit a binary generalized additive mixed model using AGHQ with exact gradients of the
approximate log-marginal likelihood.
}
\details{
TODO
}
\seealso{
Other aghqmm: 
\code{\link{aghqmm}()},
\code{\link{aghqmm_control}()}
}
\concept{aghqmm}
