% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghqmmfunctions.R
\name{aghqmm}
\alias{aghqmm}
\title{Binary Mixed Model via AGHQ with Exact Gradients}
\usage{
aghqmm(
  formula,
  data,
  k = 5,
  family = stats::binomial(),
  method = c("lbfgs", "newton", "GLMMadaptive", "lme4", "glmmEP"),
  control = aghqmm_control()
)
}
\arguments{
\item{formula}{An R mixed model formula pf the form \code{y ~ x*t + (t|id)},
compatible with \code{lme4::glmer}; see that function for details.}

\item{data}{A \code{data.frame} containing the variables found in \code{formula}.}

\item{k}{Order of the adaptive quadrature. In 1 dimension, \code{k} is the number of
points; in d-dimensions there are \code{k^d} total points.}

\item{family}{\code{R glm} family; currently only \code{binomial()} is supported
and this assumes a binary response.}

\item{method}{Method to use to fit the model. \code{lbfgs} implements the novel
exact gradient-based quasi-Newton method from the paper; \code{newton} uses the
exact gradient to build a finite-differenced Hessian for use in Newton optimization.
The other options call others' software to facilitate comparison.}

\item{control}{Output of \code{aghqmm_control()}, see that function.}
}
\value{
A list containing elements: TODO
}
\description{
Fit a binary mixed model using AGHQ with exact gradients of the
approximate log-marginal likelihood.
}
\details{
TODO
}
\examples{
library(mlmRev)
data("Contraception")
aghqmm::aghqmm(use ~ age + urban + (age|district),data=Contraception)

\dontrun{
## Not run
library(lme4)
glmer(use ~ age + urban + (age|district),data=Contraception,family = binomial)
library(GLMMadaptive)
mixed_model(use ~ age + urban,random = ~age|district,data=Contraception,family=binomial)
library(glmmEP)
reterms <- lme4::glFormula(formula,data=data,family=stats::binomial)
X <- reterms$X
Z <- X[ ,c(1,2)]
idvec <- reterms$reTrms$flist$district
glmmEP(as.numeric(Contraception$use)-1,X,Z,idvec) # note: singular
}


}
\seealso{
Other aghqmm: 
\code{\link{aghqmm_control}()}
}
\concept{aghqmm}
