% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghqmmfunctions.R
\name{aghqmm_control}
\alias{aghqmm_control}
\title{Control options for aghqmm}
\usage{
aghqmm_control()
}
\description{
Control options for the \code{aghqmm} function, see that function.
}
\seealso{
Other aghqmm: 
\code{\link{aghqmm}()}
}
\concept{aghqmm}
