% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{flatten_bdmat}
\alias{flatten_bdmat}
\title{Flatten a block-diagonal sparse matrix}
\usage{
flatten_bdmat(Z, n, d)
}
\arguments{
\item{Z}{Block diagonal sparse matrix. Can be non-square and have different number
of rows. Must have equal number of columns.}

\item{n}{Number of rows in each block, vector}

\item{d}{Number of columns in each block, scalar}
}
\value{
A dense matrix with the appropriate structure
}
\description{
Take a n x d block diagonal matrix with blocks of dimension m x p
and flatten to a (dense) matrix of dimension n x p. Taken from R package
\code{mam} by Alex Stringer.
}
