% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghqmmfunctions.R
\name{logmarglik}
\alias{logmarglik}
\title{Evaluate the approximate log-marginal likelihood and its gradient}
\usage{
logmarglik(theta, formula, data, k = 5)
}
\arguments{
\item{theta}{Point at which to evaluate the approximate log-marginal likelihood and its gradient}

\item{formula}{An R mixed model formula pf the form \code{y ~ x*t + (t|id)},
compatible with \code{lme4::glmer}; see that function for details.}

\item{data}{A \code{data.frame} containing the variables found in \code{formula}.}

\item{k}{Order of the adaptive quadrature. In 1 dimension, \code{k} is the number of
points; in d-dimensions there are \code{k^d} total points.}
}
\description{
Internal, not for user use.
}
