% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimizeaghqscalar}
\alias{optimizeaghqscalar}
\title{Fit an AGHQ model via L-BFGS optimization}
\usage{
optimizeaghqscalar(theta, y, X, nn, ww, control)
}
\arguments{
\item{theta}{Starting value for the outer parameter, containing regression
coefficients and variance components on the log-Cholesky scale.}

\item{y}{A \code{std::vector} of length \code{m} \code{Eigen} vectors of lengths \code{m_i},
containing the responses: one vector of within-group observations per group.}

\item{X}{A \code{std::vector} of length \code{m} \code{Eigen} matrices of dimensions \code{m_i x p},
representing the design matrices for the fixed effects variables.}

\item{nn}{An \code{Eigen} vector of length \code{k} containing the univariate quadrature nodes.}

\item{ww}{An \code{Eigen} vector of length \code{k} containing the univariate quadrature weights.}

\item{control}{a \code{List} containing control arguments.}
}
\description{
This function is a C++ implementation of the exact gradient-based optimization
of the AGHQ approximate log-marginal likelihood in a binary mixed model with 
scalar random intercepts. This is a special, efficient implementation of the general procedure
for multivariate Normal random effects; see below.
}
\seealso{
Other optimizeaghqmm: 
\code{\link{optimizeaghq}()}
}
\concept{optimizeaghqmm}
