% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data from a binary mixed model}
\usage{
simulate_data(m, n, beta, S)
}
\arguments{
\item{m}{Integer, number of groups}

\item{n}{Integer number of observations per group}

\item{beta}{Vector of regression coefficients for the linear predictor}

\item{S}{Either (a) variance of single random slope, or (b) 
covariance matrix of random intercept and slope}
}
\value{
A \code{data.frame} containing the covariates, group indicators, and simulated
response.
}
\description{
Simulate data from a binary mixed model with specified random
effects (co)variance. See details.
}
\details{
When \code{S} is a scalar, it is the variance of a random intercept,
and this simulates data from a Bernoulli model with \code{lme4} formula
\code{y ~ x + (1|id)}; \code{beta} should be length 2. When \code{S} is a
\code{2x2} matrix, it is the covariance matrix of a random intercept and slope,
and this simulates data from a Bernoulli model with \code{lme4} formula
\code{y ~ x*t + (t|id)}; \code{beta} should be length 4.
The covariates \code{x} and \code{t} are a group-level
indicator which is 0 for haldf the groups and 1 for the other half (\code{x}),
and an even spacing of the interval \code{(-3,3)} with \code{n} elements (\code{t}).
}
\examples{

simulate_data(10,2,c(0,0),1)
simulate_data(10,2,c(0,0,0,0),diag(2))


}
